Update tools-and-guides.md Syncthing Ryujinx Fix
Due to how Ryujinx saves folders are named in order of game opening order, need to sync entire bis folder instead of just user saves folder. This way between devices, not only are the saves synced, but between devices the same folder names will be linked in the same order.